requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
Flask==3.0.1
Flask-Cors==4.0.0
flask-restplus==0.13.0
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.4
packaging==23.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
swagger_ui==0.1.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
